requirements.txt
click==8.0.3
graphviz==0.18.2
networkx==2.6.3
pika==1.2.0
python_igraph==0.9.8
PyYAML==6.0
spotipy==2.19.0
SQLAlchemy==1.4.27
tqdm==4.62.3
psycopg2